- Refactors ALC
